feat(ClickGUI): dynamic precision of float values (#5715)
 - Introduced dynamic precision of float values depending on the max value.
-  (ScriptAPI): Fixed the default value initialization of a float setting. It expected Float and regards 0.026 as Double, and there's no way to mark it as Float (0.026F and 0.026f won't work in scripts).

+++ src/main/kotlin/net/ccbluex/liquidbounce/script/bindings/features/ScriptSetting.kt
@@ -40,14 +40,14 @@ object ScriptSetting {
     @JvmName("float")
     fun float(value: PolyglotValue): RangedValue<Float> {
         val name = value.getMember("name").asString()
-        val default = value.getMember("default").asFloat()
+        val default = value.getMember("default").asDouble()
         val range = value.getMember("range").`as`(Array<Double>::class.java)
         val suffix = value.getMember("suffix")?.asString() ?: ""
 
         require(range.size == 2)
         return rangedValue(
             name,
-            default,
+            default.toFloat(),
             range.first().toFloat()..range.last().toFloat(),
             suffix,
             ValueType.FLOAT
